refactor: consolidate timeline row layout
