+++ packages/web/src/components/tool-call-item.tsx
@@ -4,6 +4,7 @@ import type { SandboxEvent } from "@/types/session";
 import { formatSessionEventTime } from "@/lib/time";
 import { formatToolCall } from "@/lib/tool-formatters";
 import { SlackNotifyEvent } from "./slack-notify-event";
+import { TimelineRowContent } from "./timeline-row-content";
 import {
   ChevronRightIcon,
   FileIcon,
@@ -122,16 +123,9 @@ export function ToolCallItem({ event, isExpanded, onToggle, showTime = true }: T
           }`}
         />
         <ToolIcon name={formatted.icon} />
-        <span className="min-w-0 flex-1 sm:flex sm:items-start sm:gap-2">
-          <span className="block whitespace-normal [overflow-wrap:anywhere] sm:min-w-0 sm:flex-1">
-            {formatted.toolName} {formatted.summary}
-          </span>
-          {showTime && (
-            <span className="mt-0.5 block shrink-0 text-xs text-secondary-foreground sm:ml-auto sm:mt-0">
-              {time}
-            </span>
-          )}
-        </span>
+        <TimelineRowContent time={showTime ? time : undefined}>
+          {formatted.toolName} {formatted.summary}
+        </TimelineRowContent>
       </button>
 
       {isExpanded && (
